Added more tests to increase coverage and fixed a couple bugs.
We can't inherit a class from a basestring and instantiate it, so I changed
the base class to 'object' so we can still to the type checking.

I also discovered a bug with the Django 1.3 compatibility work.  In the
incr_version method, we were using the Redis's rename function to move the key,
but in Redis 2.0 you cannot rename a volitile key.

Added some moar authors